Fix macOS (Apple Silicon) build issues (#105)
* Fixing clang build issues

* Fixed sed syntax

* Fix macOS linker issues and add C++17 to protobuf target

- Add -std=c++17 to larch-autogen for abseil headers compatibility
- Exclude AppleClang from --undefined-version linker option
- Exclude AppleClang from --copy-dt-needed-entries linker option

These fixes enable successful builds of larch-dagutil and larch-dag2dot
on macOS Apple Silicon (arm64).



* Fix for protobuf build; parallel_for and mock optimizer

* Reverted tbb parallel_for_each

* Fix for parallel_for

* Fix protobuf patch timing for parallel builds (Ninja)

The previous add_custom_target approach didn't properly block
compilation of mutation_detailed.pb.cc until after the patch
was applied, causing build failures with Ninja's parallel build.

This fix uses add_custom_command with a stamp file and
OBJECT_DEPENDS to ensure the patch runs before compilation.



---------

+++ include/larch/impl/dag_loader_impl.hpp
@@ -51,9 +51,9 @@ inline int32_t EncodeBasePB(char base) {
 inline std::string FilenameFromPath(const std::string path) {
   int beg_i = int(path.size()) - 1;
   for (; beg_i >= 0; beg_i--) {
-    if (path[beg_i] == '/') {
+    if (path[static_cast<size_t>(beg_i)] == '/') {
       beg_i += 1;
-      return path.substr(beg_i, path.size());
+      return path.substr(static_cast<size_t>(beg_i), path.size());
     }
   }
   return path.substr(0, path.size());
@@ -63,7 +63,7 @@ FileFormat InferFileFormat(std::string_view path) {
   auto filename = FilenameFromPath(std::string{path});
   auto tokens = SplitString(filename, '.');
   for (int i = int(tokens.size()) - 1; i >= 0; i--) {
-    auto& extension = tokens[i];
+    auto& extension = tokens[static_cast<size_t>(i)];
     auto it =
         std::find_if(file_extension_names.begin(), file_extension_names.end(),
                      [&extension](const std::pair<std::string, FileFormat> element) {
